+++ intelmq/lib/bot_debugger.py
@@ -14,7 +14,6 @@ Depending on the subcommand received, the class either
 """
 import time
 import json
-import logging
 from os.path import exists
 from importlib import import_module
 
@@ -30,32 +29,33 @@ class BotDebugger:
     '{"source.network": "178.72.192.0/18", "time.observation": "2017-05-12T05:23:06+00:00"}' """
 
     load_configuration = utils.load_configuration
-    logging_level = "DEBUG"
-    init_log_level = {"console": logging.DEBUG, "message": logging.WARNING, "process": logging.INFO, None: logging.INFO}
+    logging_level = None
 
     def __init__(self, runtime_configuration, bot_id, run_subcommand=None, console_type=None,
-                 dryrun=None, message_kind=None, msg=None):
+                 dryrun=None, message_kind=None, msg=None, loglevel=None):
         self.runtime_configuration = runtime_configuration
-        self.leverageLogger(level=self.init_log_level[run_subcommand])
         module = import_module(self.runtime_configuration['module'])
+
+        if loglevel:
+            self.leverageLogger(loglevel)
+        elif run_subcommand == "console":
+            self.leverageLogger("DEBUG")
+
         bot = getattr(module, 'BOT')
         if run_subcommand == "message":
             bot.init = lambda *args: None
         self.instance = bot(bot_id)
 
         if not run_subcommand:
-            self.leverageLogger(logging.DEBUG)
             self.instance.start()
         else:
             self.instance._Bot__connect_pipelines()
             if run_subcommand == "console":
                 self._console(console_type)
             elif run_subcommand == "message":
-                self.leverageLogger(logging.INFO)
                 self._message(message_kind, msg)
                 return
             elif run_subcommand == "process":
-                self.leverageLogger(logging.DEBUG)
                 self._process(dryrun, msg)
             else:
                 print("Subcommand {} not known.".format(run_subcommand))
@@ -65,7 +65,7 @@ class BotDebugger:
         for console in consoles:
             try:
                 module = import_module(console)
-            except Exception as exc:
+            except Exception:
                 pass
             else:
                 if console_type and console != console_type:
@@ -148,7 +148,7 @@ class BotDebugger:
     @staticmethod
     def load_configuration_patch(*args, ** kwargs):
         d = BotDebugger.load_configuration(*args, ** kwargs)
-        if "logging_level" in d:
+        if "logging_level" in d and BotDebugger.logging_level:
             d["logging_level"] = BotDebugger.logging_level
         return d